Refactor BusActivity to BusMainFragment
