+++ app/src/main/java/com/faciee/cti/valbastrelu/eticket/repo/ETkBusRepository.java
@@ -1,7 +1,12 @@
 package com.faciee.cti.valbastrelu.eticket.repo;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
+import androidx.loader.content.AsyncTaskLoader;
+
+import android.content.Context;
 import android.os.AsyncTask;
 
 import com.faciee.cti.valbastrelu.eticket.room.EtkRoomDB;
@@ -12,6 +17,7 @@ import com.faciee.cti.valbastrelu.eticket.room.entities.Bilet;
 import com.faciee.cti.valbastrelu.eticket.ui.common.TransportType;
 import com.faciee.cti.valbastrelu.eticket.room.entities.Tranzactie;
 import com.faciee.cti.valbastrelu.eticket.room.entities.Traseu;
+import com.faciee.cti.valbastrelu.eticket.util.DummyData;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -34,9 +40,6 @@ public class ETkBusRepository {
 	private MutableLiveData<List<Traseu>> traseeLiveData;
 	private MutableLiveData<List<Tranzactie>> tranzactiiLiveData;
 	private StatiiLiveData statiiLiveData;
-	
-	private SimpleDateFormat formatterDate = new SimpleDateFormat("dd-MMM-yyyy", Locale.ENGLISH);
-	private SimpleDateFormat formatterTime = new SimpleDateFormat("HH:mm", Locale.ENGLISH);
 	
 	public static ETkBusRepository getInstance(final EtkRoomDB database) {
 		if (INSTANCE == null) {
@@ -71,7 +74,7 @@ public class ETkBusRepository {
 	public LiveData<List<Traseu>> getLiveDataTrasee(){
 		if (traseeLiveData == null){
 			traseeLiveData = new MutableLiveData<>();
-			loadTrasee();
+			traseeLiveData.setValue(DummyData.loadTrasee());
 		}
 		return traseeLiveData;
 	}
@@ -79,90 +82,14 @@ public class ETkBusRepository {
 		statiiLiveData = new StatiiLiveData(nrTraseu);
 		return statiiLiveData.getNumeStatii();
 	}
-	
-	private void loadBilete() {
-		List<Bilet> listaBilete = new ArrayList<>();
-		listaBilete.add(new Bilet(102, true, 2,2));
-		listaBilete.add(new Bilet(7, false, 2,2));
-		listaBilete.add(new Bilet(44, false, 2,2));
-		bileteLiveData.setValue(listaBilete);
-	}
-	private void loadTrasee() {
-		List<Traseu> listaTrasee = new ArrayList<>();
-		listaTrasee.add(new Traseu(9, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(10, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(11, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(12, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(13, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(14, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(15, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(16, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(17, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(18, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(19, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(20, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(22, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(23, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(24, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(25, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(26, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(27, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(28, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(29, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(30, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(31, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(34, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(35, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(36, null, TransportType.BUS));
-		listaTrasee.add(new Traseu(105, null , TransportType.BUS));
-		traseeLiveData.setValue(listaTrasee);
-	}
-	private void loadIstorice(){
-		List<Tranzactie> listaIstorice = new ArrayList<>();
-//		try {
-//			listaIstorice.add(new Tranzactie(formatterDate.parse("12-JAN-2018"),TransportType.BUS,34, -2.00));
-//			listaIstorice.add(new Tranzactie(formatterDate.parse("13-JAN-2018"),TransportType.TBUS,102, -2.00));
-//			listaIstorice.add(new Tranzactie(formatterDate.parse("16-JAN-2018"),TransportType.TRAM,7, -2.00));
-//			listaIstorice.add(new Tranzactie(formatterDate.parse("22-JAN-2018"),TransportType.BUS,7, -2.00));
-//			listaIstorice.add(new Tranzactie(formatterDate.parse("23-JAN-2018"),TransportType.PARKING,7, -1.00));
-//			listaIstorice.add(new Tranzactie(formatterDate.parse("24-JAN-2018"),TransportType.NONE,0, +15.00));
-//			listaIstorice.add(new Tranzactie(formatterDate.parse("30-JAN-2018"),TransportType.TBUS,104, -2.00));
-//			listaIstorice.add(new Tranzactie(formatterDate.parse("02-FEB-2018"),TransportType.BUS,24, -2.00));
-//			listaIstorice.add(new Tranzactie(formatterDate.parse("10-FEB-2018"),TransportType.PARKING,7, -1.00));
-//		} catch (ParseException e) {
-//			e.printStackTrace();
-//		}
-		tranzactiiLiveData.setValue(listaIstorice);
-	}
-	private void loadStatiiTest(int nrTraseu) {
-		List<String> listaStatii = new ArrayList<>();
-		listaStatii.add("Micro 19-Cinema Dacia");
-		listaStatii.add("Otelarilor");
-		listaStatii.add("Bloc D19");
-		listaStatii.add("Sala Sporturilor");
-		listaStatii.add("Flora");
-		listaStatii.add("Stadionul Otelul");
-		listaStatii.add("Ghe. Doja");
-		listaStatii.add("Piata Energiei T");
-		listaStatii.add("Liceul 9");
-		listaStatii.add("Piata Energiei R");
-		listaStatii.add("ICFrimu");
-		listaStatii.add("George Cosbuc");
-		listaStatii.add("Posta Veche");
-		listaStatii.add("Baia Comunala");
-		listaStatii.add("Piata Centrala");
-//		statiiLiveData.setValue(listaStatii);
-	}
-	
 //	LiveData<List<Statie>> getStatieForTraseu(int nrTraseu){
 //		return statieDao.getStatiiForTraseu(nrTraseu);
 //	}
 
-
 	public void insertBilet(Bilet bilet){
 		new InsertBiletAsync(biletDao, tranzactieDao).execute(bilet);
 	}
-	
+
 	private static class InsertBiletAsync extends AsyncTask<Bilet, Void, Void>{
 		private BiletDao biletDao;
 		private TranzactieDao tranzactieDao;
