+++ app/src/main/java/com/faciee/cti/valbastrelu/eticket/ui/parking/ParkingActivity.java
@@ -1,21 +1,18 @@
 package com.faciee.cti.valbastrelu.eticket.ui.parking;
 
 import android.app.Dialog;
-import androidx.lifecycle.ViewModelProviders;
 import android.os.Bundle;
-
-import com.faciee.cti.valbastrelu.eticket.databinding.ActivityParkingBinding;
-import com.google.android.material.tabs.TabLayout;
-
-import androidx.viewpager.widget.ViewPager;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.viewpager.widget.ViewPager;
+
 import com.faciee.cti.valbastrelu.eticket.R;
+import com.faciee.cti.valbastrelu.eticket.databinding.ActivityParkingBinding;
 import com.faciee.cti.valbastrelu.eticket.main.ETicketApp;
 import com.faciee.cti.valbastrelu.eticket.room.entities.BiletP;
 import com.faciee.cti.valbastrelu.eticket.ui.common.adapters.SectionsPagerAdapter;
@@ -29,10 +26,10 @@ public class ParkingActivity extends AppCompatActivity {
 	private static final String TAG = "ParkingActivity";
 	private static final int ERROR_DIALOG_REQUEST = 9001;
 	private UpdateRecyclerViewCallback recyclerViewCallback;
-	
-	private ParkingActivityViewModel parkingActivityModel;
+
 	private ActivityParkingBinding parkingBinding;
-	
+	private ParkingActivityViewModel parkingActivityModel;
+
 	private SectionsPagerAdapter mSectionsPagerAdapter;
 	
 	@Override
@@ -40,17 +37,18 @@ public class ParkingActivity extends AppCompatActivity {
 		super.onCreate(savedInstanceState);
 		parkingBinding = ActivityParkingBinding.inflate(getLayoutInflater());
 		setContentView(parkingBinding.getRoot());
-
 		parkingBinding.fab.setOnClickListener(this::onClickFab);
 
-		parkingBinding.includeToolbar.toolbar.setTitle(R.string.menu_car);
-		setSupportActionBar(parkingBinding.includeToolbar.toolbar);
+		parkingBinding.toolbar.setTitle(R.string.menu_car);
+		setActionBar(parkingBinding.toolbar);
 		parkingBinding.viewPager.setAdapter(mSectionsPagerAdapter);
 		setupViewPager(parkingBinding.viewPager);
 		parkingBinding.tabs.setupWithViewPager(parkingBinding.viewPager);
-		parkingActivityModel = ViewModelProviders.of(this).get(ParkingActivityViewModel.class);
+		parkingActivityModel = new ViewModelProvider(this).get(ParkingActivityViewModel.class);
 	}
-	
+
+
+
 	void onClickFab(View view) {
 //		Snackbar.make(view, "Bilet adaugat", Snackbar.LENGTH_LONG)
 //				.setAction("Action", null).show();
