--- app/src/main/res/layout/activity_parking.xml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-	xmlns:app="http://schemas.android.com/apk/res-auto"
-	xmlns:tools="http://schemas.android.com/tools"
-	android:id="@+id/main_content"
-	android:layout_width="match_parent"
-	android:layout_height="match_parent"
-	android:fitsSystemWindows="true"
-	tools:context=".ui.parking.ParkingActivity">
-
-	<com.google.android.material.appbar.AppBarLayout
-		android:id="@+id/appbar"
-		android:layout_width="match_parent"
-		android:layout_height="wrap_content"
-		android:paddingTop="@dimen/appbar_padding_top"
-		android:theme="@style/EticketTheme.AppBarOverlay">
-
-		<include
-			android:id="@+id/includeToolbar"
-			layout="@layout/toolbar"/>
-
-		<com.google.android.material.tabs.TabLayout
-			android:id="@+id/tabs"
-			android:layout_width="match_parent"
-			android:layout_height="wrap_content">
-
-			<com.google.android.material.tabs.TabItem
-				android:id="@+id/tabItem"
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"
-				android:text="@string/tab_name_ticket" />
-
-			<com.google.android.material.tabs.TabItem
-				android:id="@+id/tabItem2"
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"
-				android:text="@string/tab_name_parking" />
-
-			<com.google.android.material.tabs.TabItem
-				android:id="@+id/tabItem3"
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"
-				android:text="@string/tab_name_history" />
-		</com.google.android.material.tabs.TabLayout>
-
-	</com.google.android.material.appbar.AppBarLayout>
-
-	<androidx.viewpager.widget.ViewPager
-		android:id="@+id/viewPager"
-		android:layout_width="match_parent"
-		android:layout_height="match_parent"
-		app:layout_behavior="@string/appbar_scrolling_view_behavior" />
-
-	<com.google.android.material.floatingactionbutton.FloatingActionButton
-		android:id="@+id/fab"
-		android:layout_width="wrap_content"
-		android:layout_height="wrap_content"
-		android:layout_gravity="end|bottom"
-		android:layout_margin="@dimen/fab_margin"
-		app:srcCompat="@drawable/ic_fab_nfc" />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
