+++ app/src/main/res/layout/content_eticket_main.xml
@@ -7,6 +7,6 @@
 	android:background="@drawable/backgrndv1"
 	app:layout_behavior="@string/appbar_scrolling_view_behavior"
 	tools:context=".main.ETicketMain"
-	tools:showIn="@layout/app_bar_eticket_main">
+	tools:showIn="@layout/eticket_main_containter">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
