+++ app/src/main/res/layout/eticket_main_containter.xml
@@ -20,7 +20,14 @@
 
 	</com.google.android.material.appbar.AppBarLayout>
 
-	<include layout="@layout/content_eticket_main" />
+	<androidx.fragment.app.FragmentContainerView
+		app:layout_behavior="com.google.android.material.appbar.AppBarLayout$ScrollingViewBehavior"
+		android:id="@+id/nav_host_fragment"
+		android:name="androidx.navigation.fragment.NavHostFragment"
+		android:layout_width="match_parent"
+		android:layout_height="match_parent"
+		app:defaultNavHost="true"
+		app:navGraph="@navigation/nav_graph_main" />
 
 	<com.google.android.material.floatingactionbutton.FloatingActionButton
 		android:id="@+id/fab"
