+++ app/src/main/res/layout/fragment_parking_main.xml
@@ -6,18 +6,24 @@
 	android:layout_width="match_parent"
 	android:layout_height="match_parent"
 	android:fitsSystemWindows="true"
-	tools:context="com.faciee.cti.valbastrelu.eticket.ui.bus.BusActivity">
+	tools:context=".ui.parking.ParkingActivity">
 
 	<com.google.android.material.appbar.AppBarLayout
 		android:id="@+id/appbar"
 		android:layout_width="match_parent"
 		android:layout_height="wrap_content"
 		android:paddingTop="@dimen/appbar_padding_top"
-		android:theme="@style/AppTheme.AppBarOverlay">
+		android:theme="@style/EticketTheme.AppBarOverlay">
 
-		<include
-			android:id="@+id/includeToolbar"
-			layout="@layout/toolbar"/>
+		<Toolbar
+			android:id="@+id/toolbar"
+			android:layout_width="match_parent"
+			android:layout_height="?attr/actionBarSize"
+			android:layout_weight="1"
+			android:background="?attr/colorPrimary"
+			app:layout_scrollFlags="scroll|enterAlways"
+			android:theme="@style/AppTheme.AppBarOverlay"
+			app:title="@string/app_name" />
 
 		<com.google.android.material.tabs.TabLayout
 			android:id="@+id/tabs"
@@ -25,23 +31,21 @@
 			android:layout_height="wrap_content">
 
 			<com.google.android.material.tabs.TabItem
-				android:id="@+id/tabItem"
 				android:layout_width="wrap_content"
 				android:layout_height="wrap_content"
 				android:text="@string/tab_name_ticket" />
 
 			<com.google.android.material.tabs.TabItem
-				android:id="@+id/tabItem2"
 				android:layout_width="wrap_content"
 				android:layout_height="wrap_content"
-				android:text="@string/tab_name_routes" />
+				android:text="@string/tab_name_parking" />
 
 			<com.google.android.material.tabs.TabItem
-				android:id="@+id/tabItem3"
 				android:layout_width="wrap_content"
 				android:layout_height="wrap_content"
 				android:text="@string/tab_name_history" />
 		</com.google.android.material.tabs.TabLayout>
+
 	</com.google.android.material.appbar.AppBarLayout>
 
 	<androidx.viewpager.widget.ViewPager
