+++ app/src/main/res/layout/toolbar.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.appcompat.widget.Toolbar
-	xmlns:android="http://schemas.android.com/apk/res/android"
-	xmlns:app="http://schemas.android.com/apk/res-auto"
-	android:id="@+id/toolbar"
-	android:layout_width="match_parent"
-	android:layout_height="?attr/actionBarSize"
-	android:layout_weight="1"
-	android:background="?attr/colorPrimary"
-	app:layout_scrollFlags="scroll|enterAlways"
-	android:theme="@style/AppTheme.AppBarOverlay"
-	app:title="@string/app_name">
+<Toolbar
+xmlns:android="http://schemas.android.com/apk/res/android"
+xmlns:app="http://schemas.android.com/apk/res-auto"
+android:id="@+id/toolbar"
+android:layout_width="match_parent"
+android:layout_height="?attr/actionBarSize"
+android:layout_weight="1"
+android:background="?attr/colorPrimary"
+app:layout_scrollFlags="scroll|enterAlways"
+android:theme="@style/AppTheme.AppBarOverlay"
+app:title="@string/app_name">
 
-</androidx.appcompat.widget.Toolbar>
+</Toolbar>
