+++ app/src/main/res/navigation/nav_graph_main.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+	xmlns:app="http://schemas.android.com/apk/res-auto"
+	xmlns:tools="http://schemas.android.com/tools"
+	android:id="@+id/nav_graph_main"
+	app:startDestination="@id/busMainFragment">
+
+	<fragment
+		android:id="@+id/busMainFragment"
+		android:name="com.faciee.cti.valbastrelu.eticket.ui.bus.BusMainFragment"
+		android:label="fragment_bus_main"
+		tools:layout="@layout/fragment_bus_main" />
+</navigation>
